Build fixes

+++ src/libaacs/mkb.h
@@ -17,13 +17,13 @@ void mkb_close(MKB *mkb);           // free MKB
 
 uint8_t mkb_type(MKB *mkb);                                      // returns type
 uint32_t mkb_version(MKB *mkb);                                  // returns version
-uint8_t *mkb_host_revokation_entries(MKB *mkb, uint32_t *len);   // returns list of revoked host entries
-uint8_t *mkb_drive_revokation_entries(MKB *mkb, uint32_t *len);  // returns list of revoked drive entries 
-uint8_t *mkb_subdiff_records(MKB *mkb, uint32_t *len);           // returns subset-diff records
-uint8_t *mkb_subdiff_offsets(MKB *mkb, uint32_t *len);           // returns list of subset-diff offsets
-uint8_t *mkb_cvalues(MKB *mkb, uint32_t *len);                   // returns list of cvalues
+uint8_t *mkb_host_revokation_entries(MKB *mkb, size_t *len);   // returns list of revoked host entries
+uint8_t *mkb_drive_revokation_entries(MKB *mkb, size_t *len);  // returns list of revoked drive entries
+uint8_t *mkb_subdiff_records(MKB *mkb, size_t *len);           // returns subset-diff records
+uint8_t *mkb_subdiff_offsets(MKB *mkb, size_t *len);           // returns list of subset-diff offsets
+uint8_t *mkb_cvalues(MKB *mkb, size_t *len);                   // returns list of cvalues
 uint8_t *mkb_mk_dv(MKB *mkb);                                    // media key verification data return in param (always 16 bytes)
-uint8_t *mkb_signature(MKB *mkb, uint32_t *len);                 // returns MKB signature
+uint8_t *mkb_signature(MKB *mkb, size_t *len);                 // returns MKB signature
 
 /* The Media Key block is of variable size but must be a multiple of 4
  * MKB Structure:
